% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunction.R
\name{dataNormalization}
\alias{dataNormalization}
\title{dataNormalization}
\usage{
dataNormalization(small_counts, verbose, normMethod = "lib")
}
\arguments{
\item{small_counts}{Expressed matrix}

\item{normMethod}{The normalization method, 'lib'(default)}

\item{verbos}{Whether to print results}
}
\value{
Return the normalization expressed matrix
}
\description{
dataNormalization
}
