% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunction.R
\name{evaluatte_objective}
\alias{evaluatte_objective}
\title{evaluatte_objective}
\usage{
evaluatte_objective(data, sparse.matrix)
}
\arguments{
\item{data}{(genes) by column (cells) log-normalized expression matrix}

\item{sparse.matrix}{sparse matrix}
}
\value{
Return a matrix contains values of objective functions.
}
\description{
Evaluate objective function of each population
}
