% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunction.R
\name{gene_set_selection}
\alias{gene_set_selection}
\title{gene_set_selection}
\usage{
gene_set_selection(sc.counts, set_num_list = NULL)
}
\arguments{
\item{sc.counts}{Row (genes) by column (cells) log-normalized expression matrix}

\item{set_num_list}{A list provided the number of selected top HVGs for each gene subset}
}
\value{
Return a list contains logNorm expression matrix and geneSet.
}
\description{
Construct multiple gene subsets with some top HVGs
}
