% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunction.R
\name{genetic_operator_with_multi_gene}
\alias{genetic_operator_with_multi_gene}
\title{genetic_operator_with_multi_gene}
\usage{
genetic_operator_with_multi_gene(
  parent_variables,
  data,
  num.M,
  num.genes,
  num.sample,
  num.neighbor,
  delta,
  echo,
  l_limit,
  u_limit,
  crossover.p,
  knn.list,
  paralle = FALSE,
  cores = 1,
  neighbor.method = "similarity"
)
}
\arguments{
\item{parent_variables}{a list contains parent population matrices and objective values vector}

\item{data}{expressed matrix}

\item{num.M}{Number of multi-objective}

\item{num.genes}{Number of genes}

\item{num.sample}{Number of samples}

\item{delta}{Parameter delta}

\item{echo}{Parameter echo}

\item{l_limit}{Lower limitation of each genetic_operator}

\item{u_limit}{Upper limitation of each genetic_operator}

\item{crossover.p}{Probability perform crossover}

\item{paralle}{Whether implement parallel operation, 'FALSE'(default)}

\item{cores}{Number of cores, cores=1(default)}

\item{knn_id}{the k neighbors of each cell}

\item{neighbor.methdod}{Method of construct similarity matrix, 'similarity'(default)}
}
\value{
Return a pop_size by (genes+multiobjective) populaition matrix
}
\description{
genetic_operator_with_multi_gene
}
