% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunction.R
\name{initialize_variables_multi_gene}
\alias{initialize_variables_multi_gene}
\title{initialize_variables_multi_gene}
\usage{
initialize_variables_multi_gene(
  data,
  counts.list,
  num.pop,
  num.M,
  num.sample,
  min_range,
  max_range,
  knn.list,
  paralle = FALSE,
  cores = 1,
  neighbor.method = "similarity"
)
}
\arguments{
\item{data}{(genes) by column (cells) log-normalized expression matrix}

\item{counts.list}{A list geneSet contains gene's name in each gene subset}

\item{num.pop}{Number of populations}

\item{num.M}{Number of multiobjective}

\item{num.sample}{Number of cells}

\item{min_range}{lower range of each variable}

\item{max_range}{upper range of each variable}

\item{paralle}{Whether implement parallel operation, 'FALSE'(default)}

\item{cores}{Number of cores, cores=1(default)}

\item{knn_id}{the k neighbors of each cell}

\item{neighbor.methdod}{Method of construct similarity matrix, 'similarity'(default)}
}
\value{
Return a list contains multiple similarity matrices.
}
\description{
Population initialization with multiple similarity matrices
}
