% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunction.R
\name{logTransform}
\alias{logTransform}
\title{logTransform}
\usage{
logTransform(small_counts)
}
\arguments{
\item{small_counts}{Expressed matrix}
}
\value{
Return the log-normalization expressed matrix
}
\description{
logTransform
}
