% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunction.R
\name{non_domination_sort_mod}
\alias{non_domination_sort_mod}
\title{non_domination_sort_mod}
\usage{
non_domination_sort_mod(variables, num.M, num.sample, num.pop)
}
\arguments{
\item{variables}{a list contains population matrices and objective values vector}

\item{num.M}{number of multi-objective}

\item{num.sample}{number of samples}

\item{num.pop}{number of population}
}
\value{
Return a a list contains population matrices and objective values vector
}
\description{
non_domination_sort_mod
}
