% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunction.R
\name{PF_select}
\alias{PF_select}
\title{PF_select}
\usage{
PF_select(variables, alpha, only.one = FALSE)
}
\arguments{
\item{variables}{Pareto front population}

\item{alpha}{Parameter alpha}
}
\value{
Return the index of the optimal in Pareto front.
}
\description{
select the optimal from Pareto front
}
