% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunction.R
\name{predict_matrix}
\alias{predict_matrix}
\title{predict_matrix}
\usage{
predict_matrix(variables, data, index)
}
\arguments{
\item{variables}{a list contains sparse population matrices and objective values vector}

\item{data}{Expressed matrix}

\item{index}{The index of the optimal}
}
\value{
Return the imputed expressed matrix
}
\description{
predict_matrix
}
