% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunction.R
\name{regression_multi_gene}
\alias{regression_multi_gene}
\title{regression_multi_gene}
\usage{
regression_multi_gene(counts.list, neighborNum, num.sample, cores = 5)
}
\arguments{
\item{counts.list}{A list geneSet contains gene's name in each gene subset}

\item{neighborNum}{Number of nearest neighboors}

\item{num.sample}{Number of samples in regression}

\item{cores}{Number of cores}
}
\value{
Return a list contains multiple similarity matrices.
}
\description{
Construct multiple similarity matrices with regression of multiple gene subsets
}
