% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunction.R
\name{replace_chromosome}
\alias{replace_chromosome}
\title{replace_chromosome}
\usage{
replace_chromosome(intermediate_variables, num.M, num.sample, num.pop)
}
\arguments{
\item{intermediate_variables}{a list contains sparse population matrices and objective values vector}

\item{num.M}{Number of multi-objective}

\item{num.sample}{Number of samples}

\item{num.pop}{Number of populations}
}
\value{
Return a list contains sparse population matrices and objective values vector
}
\description{
replace populations to obtain Pareto optimals
}
