% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunction.R
\name{similarity_multi_gene}
\alias{similarity_multi_gene}
\title{similarity_multi_gene}
\usage{
similarity_multi_gene(counts.list, neighborNum, cores = 5)
}
\arguments{
\item{counts.list}{A list geneSet contains gene's name in each gene subset}

\item{neighborNum}{Number of nearest neighboors}

\item{cores}{Number of cores}
}
\value{
Return a list contains multiple similarity matrices.
}
\description{
Construct multiple similarity matrices with KNN graphs of multiple gene subsets
}
