% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_imputation.R
\name{sparse_imputation_with_selected_genes}
\alias{sparse_imputation_with_selected_genes}
\title{sparse_imputation_with_selected_genes}
\usage{
sparse_imputation_with_selected_genes(
  data.path = NULL,
  data,
  verbose = TRUE,
  processing = TRUE,
  evalutation = FALSE,
  expressed_genes_rate_low = NULL,
  expressed_genes_rate_up = NULL,
  expressed_cells_rate_low = NULL,
  expressed_cells_rate_up = NULL,
  subset_genes = NULL,
  num.pop = 30,
  num.Iteration = 30,
  num.M = 2,
  mu = 5,
  mum = 20,
  crossover.p = 0.5,
  set_num_list = c(50, 250, 500, 800, 1000),
  K = 50,
  delta = 0.5,
  echo = 0.7,
  paralle = FALSE,
  cores = 1,
  neighbor.method = "similarity"
)
}
\arguments{
\item{data.path}{the path of the data sets}

\item{data}{Row (genes) by column (cells) log-normalized expression matrix}

\item{verbose}{Whether to print results}

\item{processing}{Whether to process data}

\item{evalutation}{Whether to evaluate the results}

\item{expressed_genes_rate_low}{the filtering condition of low expressed genes}

\item{expressed_genes_rate_up}{the filtering condition of high expressed genes}

\item{expressed_cells_rate_low}{the filtering condition of low expressed cells}

\item{expressed_cells_rate_up}{the filtering condition of high expressed cells}

\item{subset_genes}{Number of genes}

\item{num.pop}{Number of populations}

\item{num.Iteration}{Number of iterations}

\item{num.M}{Number of multi-objective}

\item{mu}{Parameter mu}

\item{mum}{Parameter mum}

\item{crossover.p}{Probability perform crossover}

\item{set_num_list}{A list provided the number of selected top HVGs for each gene subset}

\item{K}{Number of nearest neighbors}

\item{delta}{Parameter delta}

\item{echo}{Parameter echo}

\item{paralle}{Whether implement parallel operation, 'FALSE'(default)}

\item{cores}{Number of cores}

\item{neighbor.method}{Method of construct similarity matrix, 'similarity'(default)}
}
\value{
Return a list contains logNorm expression matrix and geneSet.
}
\description{
sparse imputation with nearest neighbors
}
