% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunction.R
\name{SumoneNormalization}
\alias{SumoneNormalization}
\title{SumoneNormalization}
\usage{
SumoneNormalization(data)
}
\arguments{
\item{data}{data matrix}
}
\value{
Return a sum normalized matrix
}
\description{
sum normalization
}
