% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunction.R
\name{tournament_selection}
\alias{tournament_selection}
\title{tournament_selection}
\usage{
tournament_selection(variables, num.M, num.pop, pool_size, tour_size)
}
\arguments{
\item{variables}{a list contains sparse population matrices and objective values vector}

\item{num.M}{Number of multi-objective}

\item{num.pop}{Number of populations}

\item{pool_size}{Number of populations used for pooling}

\item{tour_size}{Number of populations used for tour}
}
\value{
Return a list contains sparse population matrices and objective values vector
}
\description{
tournament selection to obtain Pareto optimals
}
